+++ gateway.tf
@@ -143,6 +143,7 @@ resource "aws_api_gateway_method_response" "isolation_get_method_response" {
     "method.response.header.Cache-Control"             = true,
     "method.response.header.Pragma"                    = true,
     "method.response.header.Strict-Transport-Security" = true
+    "method.response.header.X-Frame-Options"           = true
   }
 }
 
@@ -227,6 +228,223 @@ resource "aws_api_gateway_integration_response" "isolation_key_get_integration_r
   }
 }
 
+## /admin-ui
+resource "aws_api_gateway_resource" "admin_ui" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_rest_api.main.root_resource_id
+  path_part   = "admin-ui"
+}
+resource "aws_api_gateway_method" "admin_ui_get" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.admin_ui.id
+  http_method      = "GET"
+  authorization    = "NONE"
+  api_key_required = false
+}
+
+resource "aws_api_gateway_integration" "admin_ui_get_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.admin_ui.id
+  http_method             = aws_api_gateway_method.admin_ui_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
+  integration_http_method = "GET"
+  type                    = "AWS"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/admin-ui/index.html", var.aws_region, aws_s3_bucket.assets.id)
+  credentials             = aws_iam_role.gateway.arn
+}
+
+resource "aws_api_gateway_method_response" "admin_ui_get" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.admin_ui.id
+  http_method = aws_api_gateway_method.admin_ui_get.http_method
+  status_code = "200"
+
+  response_parameters = {
+    "method.response.header.Content-Length"            = true,
+    "method.response.header.Content-Type"              = true,
+    "method.response.header.Cache-Control"             = true,
+    "method.response.header.Pragma"                    = true,
+    "method.response.header.Strict-Transport-Security" = true
+  }
+}
+
+resource "aws_api_gateway_integration_response" "admin_ui_get_integration" {
+  rest_api_id       = aws_api_gateway_rest_api.main.id
+  resource_id       = aws_api_gateway_resource.admin_ui.id
+  http_method       = aws_api_gateway_method.admin_ui_get.http_method
+  selection_pattern = aws_api_gateway_method_response.admin_ui_get.status_code
+  status_code       = aws_api_gateway_method_response.admin_ui_get.status_code
+  response_parameters = {
+    "method.response.header.Content-Length"            = "integration.response.header.Content-Length",
+    "method.response.header.Content-Type"              = "integration.response.header.Content-Type",
+    "method.response.header.Cache-Control"             = "'no-store'",
+    "method.response.header.Pragma"                    = "'no-cache'",
+    "method.response.header.Strict-Transport-Security" = format("'max-age=%s; includeSubDomains'", var.hsts_max_age)
+  }
+}
+
+## /admin-ui/{key+}
+
+resource "aws_api_gateway_resource" "admin_ui_key" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_resource.admin_ui.id
+  path_part   = "{key+}"
+}
+
+resource "aws_api_gateway_method" "admin_ui_key_get" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.admin_ui_key.id
+  http_method      = "GET"
+  authorization    = "NONE"
+  api_key_required = false
+  request_parameters = {
+    "method.request.path.key" = true
+  }
+}
+resource "aws_api_gateway_integration" "admin_ui_key_get_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.admin_ui_key.id
+  http_method             = aws_api_gateway_method.admin_ui_key_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
+  integration_http_method = "GET"
+  type                    = "AWS"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/admin-ui/{key}", var.aws_region, aws_s3_bucket.assets.id)
+  credentials             = aws_iam_role.gateway.arn
+  request_parameters = {
+    "integration.request.path.key" = "method.request.path.key",
+  }
+}
+
+resource "aws_api_gateway_method_response" "admin_ui_key_get_method_response" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.admin_ui_key.id
+  http_method = aws_api_gateway_method.admin_ui_key_get.http_method
+  status_code = "200"
+  response_models = {
+    "application/json" = "Empty"
+  }
+  response_parameters = {
+    "method.response.header.Content-Length"            = false,
+    "method.response.header.Content-Type"              = false,
+    "method.response.header.Cache-Control"             = true,
+    "method.response.header.Pragma"                    = true,
+    "method.response.header.Strict-Transport-Security" = true
+  }
+}
+
+resource "aws_api_gateway_integration_response" "admin_ui_key_get_integration_response" {
+  rest_api_id       = aws_api_gateway_rest_api.main.id
+  resource_id       = aws_api_gateway_resource.admin_ui_key.id
+  http_method       = aws_api_gateway_method.admin_ui_key_get.http_method
+  selection_pattern = aws_api_gateway_method_response.admin_ui_key_get_method_response.status_code
+  status_code       = aws_api_gateway_method_response.admin_ui_key_get_method_response.status_code
+  response_parameters = {
+    "method.response.header.Content-Length"            = "integration.response.header.Content-Length",
+    "method.response.header.Content-Type"              = "integration.response.header.Content-Type",
+    "method.response.header.Cache-Control"             = "'no-store'",
+    "method.response.header.Pragma"                    = "'no-cache'",
+    "method.response.header.Strict-Transport-Security" = format("'max-age=%s; includeSubDomains'", var.hsts_max_age)
+  }
+}
+
+## /admin
+resource "aws_api_gateway_resource" "admin" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_rest_api.main.root_resource_id
+  path_part   = "admin"
+}
+
+resource "aws_api_gateway_authorizer" "admin_authorizer" {
+  name          = "${module.labels.id}-cognito-authorizer"
+  rest_api_id   = aws_api_gateway_rest_api.main.id
+  type          = "COGNITO_USER_POOLS"
+  provider_arns = [aws_cognito_user_pool.admin_user_pool.arn]
+}
+
+
+## /admin/{proxy}
+resource "aws_api_gateway_resource" "admin_proxy" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_resource.admin.id
+  path_part   = "{proxy+}"
+}
+
+resource "aws_api_gateway_method" "admin_proxy_options" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.admin_proxy.id
+  http_method      = "OPTIONS"
+  authorization    = "NONE"
+  api_key_required = false
+  request_parameters = {
+    "method.request.path.proxy" = true
+  }
+}
+
+resource "aws_api_gateway_integration" "admin_proxy_options_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.admin_proxy.id
+  http_method             = aws_api_gateway_method.admin_proxy_options.http_method
+  integration_http_method = "OPTIONS"
+  type                    = "HTTP_PROXY"
+  uri                     = format("http://%s/{proxy}", aws_lb.admin.dns_name)
+
+  request_parameters = {
+    "integration.request.path.proxy"              = "method.request.path.proxy",
+    "integration.request.header.X-Routing-Secret" = "'${jsondecode(data.aws_secretsmanager_secret_version.api_gateway_header.secret_string)["header-secret"]}'",
+    "integration.request.header.X-Forwarded-For"  = "'nope'"
+  }
+}
+
+resource "aws_api_gateway_method" "admin_proxy_any" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.admin_proxy.id
+  http_method      = "ANY"
+  authorization    = "COGNITO_USER_POOLS"
+  authorizer_id    = aws_api_gateway_authorizer.admin_authorizer.id
+  api_key_required = false
+  request_parameters = {
+    "method.request.path.proxy" = true
+  }
+}
+
+resource "aws_api_gateway_integration" "admin_proxy_any_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.admin_proxy.id
+  http_method             = aws_api_gateway_method.admin_proxy_any.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
+  integration_http_method = "ANY"
+  type                    = "HTTP_PROXY"
+  uri                     = format("http://%s/{proxy}", aws_lb.admin.dns_name)
+  request_parameters = {
+    "integration.request.path.proxy"              = "method.request.path.proxy",
+    "integration.request.header.X-Routing-Secret" = "'${jsondecode(data.aws_secretsmanager_secret_version.api_gateway_header.secret_string)["header-secret"]}'",
+    "integration.request.header.X-Forwarded-For"  = "'nope'"
+  }
+}
+
+resource "aws_api_gateway_method_response" "admin_proxy_any" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.admin_proxy.id
+  http_method = aws_api_gateway_method.admin_proxy_any.http_method
+  status_code = "200"
+
+  response_parameters = {
+    "method.response.header.access-control-allow-headers" = true,
+    "method.response.header.access-control-allow-methods" = true,
+    "method.response.header.access-control-allow-origin"  = true
+  }
+}
+
+resource "aws_api_gateway_integration_response" "admin_proxy_any_integration" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.admin_proxy.id
+  http_method = aws_api_gateway_method.admin_proxy_any.http_method
+  status_code = aws_api_gateway_method_response.admin_proxy_any.status_code
+
+  depends_on = [ 
+    aws_api_gateway_integration.admin_proxy_any_integration
+   ]
+}
 
 ## /api
 resource "aws_api_gateway_resource" "api" {
@@ -766,6 +984,10 @@ resource "aws_api_gateway_deployment" "live" {
     aws_api_gateway_integration.root,
     aws_api_gateway_integration.isolation_get_integration,
     aws_api_gateway_integration.isolation_key_get_integration,
+    aws_api_gateway_integration.admin_ui_get_integration,
+    aws_api_gateway_integration.admin_ui_key_get_integration,
+    aws_api_gateway_integration.admin_proxy_options_integration,
+    aws_api_gateway_integration.admin_proxy_any_integration,
     aws_api_gateway_integration.api_proxy_options_integration,
     aws_api_gateway_integration.api_proxy_any_integration,
     aws_api_gateway_integration.api_settings_get_integration,
@@ -829,4 +1051,16 @@ resource "aws_api_gateway_authorizer" "main" {
   rest_api_id            = aws_api_gateway_rest_api.main.id
   authorizer_uri         = coalesce(join("", aws_lambda_alias.authorizer_live.*.invoke_arn), aws_lambda_function.authorizer.invoke_arn)
   authorizer_credentials = aws_iam_role.authorizer.arn
+}
+
+# #########################################
+# Default Responses
+# #########################################
+resource "aws_api_gateway_gateway_response" "test" {
+  rest_api_id   = aws_api_gateway_rest_api.main.id
+  response_type = "DEFAULT_4XX"
+
+  response_parameters = {
+    "gatewayresponse.header.access-control-allow-origin" = "'*'"
+  }
 }