+++ lambda-callback.tf
@@ -21,7 +21,8 @@ data "aws_iam_policy_document" "callback_policy" {
         aws_ssm_parameter.db_database.arn,
         aws_ssm_parameter.db_host.arn,
         aws_ssm_parameter.db_port.arn,
-        aws_ssm_parameter.db_ssl.arn
+        aws_ssm_parameter.db_ssl.arn,
+        aws_ssm_parameter.time_zone.arn
       ],
       aws_ssm_parameter.callback_email_notifications_sns_arn.*.arn
     )