+++ lambda-download.tf
@@ -10,13 +10,17 @@ module "download" {
   enable = contains(var.optional_lambdas_to_include, "download")
   name   = format("%s-download", module.labels.id)
 
-  aws_parameter_arns = [
+  aws_parameter_arns = concat([
     aws_ssm_parameter.db_database.arn,
     aws_ssm_parameter.db_host.arn,
     aws_ssm_parameter.db_port.arn,
     aws_ssm_parameter.db_reader_host.arn,
-    aws_ssm_parameter.db_ssl.arn
-  ]
+    aws_ssm_parameter.db_ssl.arn,
+    aws_ssm_parameter.time_zone.arn
+    ],
+    aws_ssm_parameter.interop_origin.*.arn
+  )
+
   aws_secret_arns                = concat([data.aws_secretsmanager_secret_version.rds_read_write.arn], data.aws_secretsmanager_secret_version.interop.*.arn)
   cloudwatch_schedule_expression = var.download_schedule
   config_var_prefix              = local.config_var_prefix