+++ users.tf
@@ -34,6 +34,15 @@ data "aws_iam_policy_document" "ci_user" {
       format("%s/*", aws_s3_bucket.assets.arn)
     ]
   }
+
+  statement {
+    actions = [
+      "route53:GetChange"
+    ]
+    resources = [
+      "*"
+    ]
+  }
 }
 
 data "aws_iam_policy_document" "ci_user_lambda" {
@@ -59,6 +68,8 @@ data "aws_iam_policy_document" "ci_user_pass_role" {
     ]
 
     resources = [
+      aws_iam_role.admin_ecs_task_role.arn,
+      aws_iam_role.admin_ecs_task_execution.arn,
       aws_iam_role.api_ecs_task_role.arn,
       aws_iam_role.api_ecs_task_execution.arn,
       aws_iam_role.push_ecs_task_role.arn,